requirements.txt
boto3==1.34.57
fastapi==0.110.0
uvicorn==0.27.1
requests==2.31.0
python-dotenv==1.0.1
botocore==1.34.57
ipython==8.22.2
pydantic==2.6.3